+++ changelog.md
@@ -5,6 +5,7 @@
 ### Non-Breaking
 
 * Updated compat bounds for OrderedCollections to include 2.0
+* Added iteration over TimeSeries
 
 ## v1.0.0